requirements.txt
default

asgiref==3.7.2
beautifulsoup4==4.12.2
distlib==0.3.7
Django==4.2.5
django-bootstrap-v5==1.0.11
django-phonenumber-field==7.2.0
filelock==3.13.1
md5hash==0.0.6
mysqlclient==2.2.0
phonenumbers==8.13.24
phonenumberslite==8.13.24
platformdirs==3.11.0
PyMySQL==1.1.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6


entorno env

asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
django-phonenumber-field==7.2.0
md5hash==0.0.6
mysqlclient==2.2.0
phonenumbers==8.13.25
phonenumberslite==8.13.25
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3




